refactor: organize project structure more
