+++ playground/src/Playground.tsx
@@ -2,8 +2,7 @@ import { useState } from "react";
 
 import { Header } from "./components/Header";
 import { MainLayout } from "./components/MainLayout";
-import { OutputEditor } from "./components/editor/OutputEditor";
-import { SourceEditor } from "./components/editor/SourceEditor";
+import { OutputEditor, SourceEditor } from "./components/editor";
 import { FormatOptionsContent } from "./components/forms/SettingsPanel";
 import { DEFAULT_FORMAT_OPTIONS } from "./constants";
 import { useInitialSample, useScreenSize, useTypstFormatter } from "./hooks";
