+++ playground/src/components/editor/index.ts
@@ -1,0 +1,3 @@
+export { CodeEditor } from "./CodeEditor";
+export { OutputEditor } from "./OutputEditor";
+export { SourceEditor } from "./SourceEditor";
