+++ playground/src/contexts/index.ts
@@ -1,2 +1,1 @@
-export { ThemeProvider } from "./ThemeContext";
-export { useTheme } from "./useTheme";
+export { ThemeProvider } from "./ThemeProvider";
